rework to not use rule34.xxx api
the rule34.xxx api now requires an api key and is restricted to 50
requests per second. sadly the api still doesn't feature a way to access
a users favorites and so requires an api request per favorited post.

since using a userscript is still required to fetch the favorites pages
we might as well switch to using the userscript to fetch the image data
as well.

in my testing allows us to roughly double the download speed.